Reorganize core-components
