+++ packages/core-components/src/components/ContainedContent.vue
@@ -26,7 +26,7 @@ import type { ContentElement } from '@tailor-cms/interfaces/content-element';
 import type { ContentElementCategory } from '@tailor-cms/interfaces/schema';
 import { get, throttle } from 'lodash-es';
 
-import ContentElementWrapper from './ContentElement.vue';
+import ContentElementWrapper from './ContentElement/index.vue';
 
 interface Props {
   element: ContentElement;
