+++ packages/core-components/src/components/ContentElement/ElementActions/ElementLinkedIndicator.vue
