+++ packages/core-components/src/components/ContentElement/ElementActions/index.vue
