+++ packages/core-components/src/index.ts
@@ -1,6 +1,6 @@
 export { default as CircularProgress } from './components/CircularProgress.vue';
 export { default as ContainedContent } from './components/ContainedContent.vue';
-export { default as ContentElement } from './components/ContentElement.vue';
+export { default as ContentElement } from './components/ContentElement/index.vue';
 export { default as Discussion } from './components/Discussion/index.vue';
 export { default as ElementList } from './components/ElementList.vue';
 export { default as AddElement } from './components/AddElement/index.vue';
@@ -16,7 +16,8 @@ export { default as AssetInput } from './components/AssetInput.vue';
 export { default as ElementPlaceholder } from './components/ElementPlaceholder.vue';
 export { default as PreviewOverlay } from './components/PreviewOverlay.vue';
 export { default as RadarChart } from './components/RadarChart.vue';
-export { default as QuestionContainer } from './components/QuestionContainer/index.vue';
+export { default as QuestionContainer }
+  from './components/QuestionElement/QuestionContainer/index.vue';
 export { default as EmbeddedContainer } from './components/EmbeddedContainer.vue';
 export { default as TailorDialog } from './components/TailorDialog.vue';
 export { default as TailorEmptyState } from './components/TailorEmptyState.vue';
